+++ imports/ui/components/company-profile/tabs/jobs.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import i18n from "meteor/universe:i18n";
 
-import JobPosting from "./jobPosting.jsx";
+import JobPosting from "/imports/ui/components/jobPosting.jsx";
 
 const T = i18n.createComponent();
