+++ imports/ui/components/company-profile/tabs/reviews.jsx
@@ -4,8 +4,8 @@ import i18n from "meteor/universe:i18n";
 
 import { urlGenerators } from "/imports/startup/client/router.jsx";
 
-import CompanyReview from "./companyReview.jsx";
-import CompanyRating from "./companyRatingsComponent.jsx";
+import CompanyReview from "/imports/ui/components/companyReview.jsx";
+import CompanyRating from "/imports/ui/components/companyRatingsComponent.jsx";
 
 const T = i18n.createComponent();
